backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.10.1
plotly==5.16.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
aif360==0.5.0
simple-salesforce==1.12.4